% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_rmds.R
\name{get_rmds}
\alias{get_rmds}
\title{gets all rmd files in a directory to render}
\usage{
get_rmds(dir, exclude = NULL, recursive = TRUE)
}
\arguments{
\item{dir}{directory to search}

\item{exclude}{vector of rmarkdown doc names to exclude}

\item{recursive}{whether to search recursively default to TRUE}
}
\value{
vector of Rmd files with full normalized paths
}
\description{
gets all rmd files in a directory to render
}

