% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rendering.R
\name{prepare_book_dir}
\alias{prepare_book_dir}
\title{copy all files to temp dir to prepare to render}
\usage{
prepare_book_dir(dir)
}
\arguments{
\item{dir}{}
}
\value{
path of book directory for rendering
}
\description{
copy all files to temp dir to prepare to render
}

